The most significant changes involve the implementation of the repository pattern and the unit of work pattern, which are used to abstract the data access layer and handle transaction-related operations, respectively. This is reflected in the creation of new interfaces and classes, such as `IRepository.cs`, `IUnitOfWork.cs`, `Repository.cs`, and `UnitOfWork.cs`. The `VillaFracaoRepository.cs` and `VillaRepository.cs` classes were also created to handle operations related to VillaFracao and Villa entities.
Changes were also made to the `Program.cs` file to include using statements for the new namespaces and to add the `IUnitOfWork` service to the service container. The `IVillaRepository` service was commented out.

The `Explicacao.md` file was updated with explanations and instructions on various topics, including how to add images, work with models, use tag helpers, and implement dependency injection.

The `Apagar.cshtml`, `Atualizar.cshtml`, and `Create.cshtml` files were modified to improve the form fields and layout. The `Index.cshtml` file was updated to improve the user interface.

Lastly, the `Apagar.cshtml` file was updated to include a form for deleting a `VillaFracao` instance.